Added to android project. Broken up to specific clients depending on type of data. Changed tests for how server is supposed to work now
