+++ android/app/src/main/java/com/example/aida/socketcommunication/SocketHelpers.kt
@@ -1,4 +1,4 @@
-package org.example
+package com.example.aida.socketcommunication
 
 import java.io.File
 import javax.imageio.ImageIO
