+++ android/app/src/main/java/com/example/aida/socketcommunication/socketbild.jpeg
